Dynamic hash plugin: save on CPU cycles
